[GREEN-67] return updated Booking DTO
